# docs/swagger.yaml
basePath: /
definitions:
...
  models.Artifact:
    properties:
      ID:
        type: string
      ImageURL:
        type: string
      Status:
        type: string
      description:
        type: string
      end_date:
        type: integer
      epoch:
        type: string
      name:
        type: string
      start_date:
        type: integer
      tpq:
        type: integer
    type: object
  models.TPQRequest:
    properties:
      completedAt:
        type: string
      createdAt:
        type: string
      creatorID:
        type: integer
      excavation:
        type: string
      formedAt:
        type: string
      id:
        type: string
      moderatorID:
        type: integer
      result:
        description: Изменено на *int для nullable (пустое значение)
        type: integer
      status:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API for managing artifacts and TPQ requests.
  title: Chronus API
  version: "1.0"
paths:
...
  /api/artifacts:
    get:
      consumes:
      - application/json
      description: Get all active artifacts with optional filter
      parameters:
      - description: Filter by name or TPQ
        in: query
        name: filter
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Artifact'
            type: array
      summary: Get list of artifacts
      tags:
      - artifacts
    post:
      consumes:
      - application/json
      description: Create a new artifact (user required)
      parameters:
      - description: Artifact data
        in: body
        name: artifact
        required: true
        schema:
          $ref: '#/definitions/models.Artifact'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Artifact'
        "400":
          description: Invalid request body
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create new artifact
      tags:
      - artifacts
  /api/artifacts/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an artifact (user required)
      parameters:
      - description: Artifact ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Artifact not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete artifact
      tags:
      - artifacts
    get:
      consumes:
      - application/json
      description: Get details of a specific artifact
      parameters:
      - description: Artifact ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Artifact'
        "404":
          description: Artifact not found
          schema:
            type: string
      summary: Get artifact by ID
      tags:
      - artifacts
    put:
      consumes:
      - application/json
      description: Update an existing artifact (user required)
      parameters:
      - description: Artifact ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated artifact data
        in: body
        name: artifact
        required: true
        schema:
          $ref: '#/definitions/models.Artifact'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Artifact'
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Artifact not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update artifact
      tags:
      - artifacts
  /api/artifacts/{id}/add_to_request:
    post:
      consumes:
      - application/json
      description: Add an artifact to the current draft request (user required)
      parameters:
      - description: Artifact ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TPQRequest'
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Artifact not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Add artifact to request
      tags:
      - artifacts
  /api/artifacts/{id}/image:
    post:
      consumes:
      - multipart/form-data
      description: Upload an image for the artifact (user required)
      parameters:
      - description: Artifact ID
        in: path
        name: id
        required: true
        type: string
      - description: Image file
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Artifact'
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Artifact not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Upload image for artifact
      tags:
      - artifacts
...
  /artifact/{id}:
    get:
      description: Render HTML page with artifact details
      parameters:
      - description: Artifact ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - text/html
      responses:
        "200":
          description: HTML page
          schema:
            type: string
        "404":
          description: Artifact not found
          schema:
            type: string
      summary: Display artifact details
      tags:
      - artifacts
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"